requirements.txt
beautifulsoup4==4.4.1
docutils==0.12
enum34==1.1.1
flake8==2.5.1
ipaddress==1.0.15
isort==4.2.2
natsort==4.0.4
pep8==1.6.2
pies2overrides==2.6.7
pies==2.6.7
py==1.4.31
pyflakes==1.0.0
Pygments==2.0.2
pytest-django==2.9.1
pytest==2.8.4
sqlparse==0.1.18
